店铺消息暂时提交

+++ shop-service/src/main/java/com/masiis/shop/web/mall/service/message/SfMessageSrRelationService.java
@@ -18,7 +18,7 @@ public class SfMessageSrRelationService {
     @Resource
     private SfMessageSrRelationMapper srRelationMapper;
 
-    public Integer countNumsFromUser(ComUser user) {
+    public Long countNumsFromUser(ComUser user) {
         return srRelationMapper.countNumsFromUser(user.getId());
     }
 }
